requirements.txt
fastapi
uvicorn[standard]
pydantic
psycopg2-binary
openai
pandas
boto3
sqlparse
